instrumentation/requirements.txt
fastapi
uvicorn
flask
httpx
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-system-metrics
opentelemetry-instrumentation-httpx